Threshold floor values

+++ src/core/kernel.cpp
@@ -148,10 +148,10 @@ namespace Core
         }
 
         /** G] Check the nNonce Efficiency Proportion Requirements. **/
-        double nThreshold = ((nTime - vtx[0].nTime) * 100.0) / nNonce;
-        double nRequired  = ((50.0 - nTrustWeight - nBlockWeight) * MAX_STAKE_WEIGHT) / std::min((int64)MAX_STAKE_WEIGHT, vtx[0].vout[0].nValue);
+        unsigned int nThreshold = ((nTime - vtx[0].nTime) * 100.0) / nNonce;
+        unsigned int nRequired  = ((50.0 - nTrustWeight - nBlockWeight) * MAX_STAKE_WEIGHT) / std::min((int64)MAX_STAKE_WEIGHT, vtx[0].vout[0].nValue);
         if(nThreshold < nRequired)
-            return error("CBlock::VerifyStake() : Coinstake / nNonce threshold too low %f Required %f. Energy efficiency limits Reached", nThreshold, nRequired);
+            return error("CBlock::VerifyStake() : Coinstake / nNonce threshold too low %u Required %u. Energy efficiency limits Reached", nThreshold, nRequired);
 
 
         /** H] Check the Block Hash with Weighted Hash to Target. **/
